+++ nuxt.config.ts
@@ -327,6 +327,7 @@ export default defineNuxtConfig({
                 'ol/math.js',
                 'ol/geom.js',
                 'ol/style/Circle.js',
+                'dexie',
             ],
         },
         build: {